+++ src/Definition/Exception/InvalidAnnotation.php
@@ -5,10 +5,10 @@ declare(strict_types=1);
 namespace DI\Definition\Exception;
 
 /**
- * Exception in the definitions using annotations.
+ * Error in the definitions using annotations.
  *
  * @author Matthieu Napoli <matthieu@mnapoli.fr>
  */
-class AnnotationException extends InvalidDefinition
+class InvalidAnnotation extends InvalidDefinition
 {
 }